requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
defusedxml==0.7.1
idna==3.10
pip==24.0
requests==2.32.4
urllib3==2.5.0
youtube-transcript-api==1.2.1
yt-dlp==2024.12.13
openai==1.51.0
python-dotenv==1.0.0
openai-whisper==20231117
ffmpeg-python==0.2.0
